+++ aat-lib/src/main/java/ch/bailu/aat_lib/map/layer/LayerWrapper.kt
@@ -4,7 +4,6 @@ import ch.bailu.aat_lib.map.MapContext
 import ch.bailu.aat_lib.service.ServicesInterface
 import org.mapsforge.core.graphics.Canvas
 import org.mapsforge.core.model.BoundingBox
-import org.mapsforge.core.model.LatLong
 import org.mapsforge.core.model.Point
 import org.mapsforge.core.model.Rotation
 import org.mapsforge.map.layer.Layer
@@ -16,9 +15,5 @@ class LayerWrapper(
 ) : Layer() {
     override fun draw(bounding: BoundingBox, zoom: Byte, canvas: Canvas, topLeftPoint: Point, rotation: Rotation) {
         services.insideContext { layer.drawInside(mcontext) }
-    }
-
-    override fun onTap(tapLatLong: LatLong, layerXY: Point?, tapXY: Point): Boolean {
-        return layer.onTap(ch.bailu.aat_lib.util.Point(tapXY.x, tapXY.y))
     }
 }